dataend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.103.1
google-api-core==2.12.0
google-auth==2.23.0
google-cloud-language==2.11.1
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
httptools==0.6.0
idna==3.4
JPype1==1.4.1
konlpy==0.6.0
lxml==4.9.3
numpy==1.26.0
packaging==23.1
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
unicorn==2.0.1.post1
urllib3==1.26.16
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
